Finished muting letters (#9)
Updated comments telling which mod makes a change to vanilla code
Fixed Letter for Task_Increase_Control not showing
Added the if/else need to figure out if a letter should be send, or just perform the action silently.
Added the additional text to description of tasks, to indicate that no letter will be sent.
Fixed canceling mute not removing variables
Fixed council titles not showing it mute events.